HW2/requirements.txt
mysql-connector-python
python-dotenv
pytest
pytest-mock